test delete add wine order

+++ client/src/components/WinesAdm.vue
@@ -395,9 +395,9 @@ export default {
       await this.$store.dispatch("wines/searchWinesByColor", couleur);
     },
     async Delete(name, id) {
-      // if (confirm("Attention : Vous êtes sur le point de supprimer " + name)) {
+       if (confirm("Attention : Vous êtes sur le point de supprimer " + name)) {
       await this.$store.dispatch("wines/deleteWine", id);
-      // }
+       }
     },
   },
 };
